requirements.txt
Flask==3.0.0
pdfplumber==0.11.0
streamlit==1.38.0
google-generativeai==0.8.3
pytest==8.3.2
